requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
openai==1.42.0
requests==2.31.0
Faker==13.3.4